+++ src/window.js
@@ -5,7 +5,9 @@
  *  Copyright: (c) 2020 Benny A. Daon - benny@tuzig.com
  *  License: GPLv3
  */
-import { Layout, Pane } from './cells.js'
+import { Layout } from './layout.js'
+import { Cell } from './cell.js'
+import { Pane } from './pane.js'
 import * as Hammer from 'hammerjs'
 
 const ABIT = 10
@@ -70,8 +72,8 @@ export class Window {
         this.nameE.classList.add("on")
         this.gate.activeW = this
         window.location.href=`#tab-${this.gate.id}.${this.id+1}`
-        this.gate.sendState()
-        terminal7.run(_ => this.activeP.focus(), ABIT)
+        if (this.activeP)
+            this.activeP.focus()
     }
     addLayout(dir, basedOn) {
         let l = new Layout(dir, basedOn)
@@ -103,7 +105,7 @@ export class Window {
             else {
                 let p = l.addPane(cell)
                 if (cell.active)
-                    p.focus()
+                    this.activeP = p
             }
         })
         return l
@@ -133,7 +135,7 @@ export class Window {
             terminal7.run(() => {
                 p.innerHTML = p.w.name
                 this.activeP.focus()
-            }, 0)
+            }, ABIT)
         }, { once: true })
         i.addEventListener('change', (e) => {
             console.log("change", e)
@@ -200,8 +202,9 @@ export class Window {
                 if (match(c))
                     nextPane = c
         })
-        if (nextPane)
+        if (nextPane) {
             nextPane.focus()
-            
+            this.gate.sendState()
+        }
     }
 }